aaaaaaaaaa

+++ src/conf.rs
@@ -16,7 +16,7 @@ pub const FPS: u32 = 60;
 pub const SECOND_IN_MICROSECOND: u32 = 1_000_000;
 pub const ONE_FRAME_IN_MICROSECONDS: u32 = SECOND_IN_MICROSECOND / FPS;
 
-pub const CYCLE_PER_MCYCLE: u64 = 4;
+pub const CYCLE_PER_MCYCLE: u32 = 4;
 
 // Cycles per second.
 pub const CPU_HZ: u32 = 4194304;
